pyproject.toml
[tool.poetry]
name = "fremont-app-backend"
version = "0.1.0"
description = "Backend for Fremont app"
authors = ["Mohamad <mmortada685@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
psycopg2 = "^2.9.1"
django-stubs = "^1.8.0"
django-dotenv = "^1.4.2"
rules = "^3.4"
django-allauth = "^0.63.3"
djoser = "^2.2.3"
django-better-admin-arrayfield = "^1.4.2"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110